@@ -53,7 +53,7 @@
     "vue-tsc": "^1.8.0"
   },
   "engines": {
-    "node": ">= 16.0.0",
+    "node": ">= 20.0.0",
     "npm": ">= 8.0.0"
   },
   "browserslist": [
