@@ -1,5 +1,5 @@
 /* eslint-disable no-return-assign */
-module.exports = {
+export default {
   props: {
     /**
      * Toggles the adding of a 'loading' class to the main
