@@ -14,15 +14,7 @@ module.exports = {
   plugins: [
     'html'
   ],
-  // check if imports actually resolve
-  'settings': {
-    'import/resolver': {
-      'webpack': {
-        'config': 'build/webpack.base.conf.js'
-      }
-    }
-  },
-  // add your custom rules here
+    // add your custom rules here
   'rules': {
     // don't require .vue extension when importing
     'import/extensions': ['error', 'always', {
