@@ -3,7 +3,7 @@
     <div class="keypad">
       <el-popover v-for="(button, i) in buttons" :key="i" placement="bottom-start" trigger="click" :ref="'popups-' + i">
         <div class="popper-content">
-          <remote-button-options :button="button" :index="i" :title="'Button ' + (i + 1)" :bus="bus"/>
+          <remote-button-options :button="button" :index="i" :title="'Button ' + (i + 1)" @update-button="updateButton"/>
           <div class="button frameless cancel narrow" @click="closeOptions(i)">
             <span class="icon cancel small"></span>
           </div>
@@ -30,24 +30,18 @@
 </template>
 
 <script>
-  import { createApp } from 'vue'
-  import RemoteButtonOptions from './RemoteButtonOptions'
+  import RemoteButtonOptions from './RemoteButtonOptions.vue'
   import { Controls } from '../../models/Control'
 
   export default {
     components: {RemoteButtonOptions},
     props: {
-      value: {type: Array, default: () => Controls.empty()}
-    },
-    mounted () {
-      // Vue 3 doesn't have $on, use mitt or similar for event bus
-      // For now, we'll handle events directly in child components
+      modelValue: {type: Array, default: () => Controls.empty()}
     },
     data () {
       return {
-        buttons: this.setup(this.value),
-        index: 0,
-        bus: createApp({})
+        buttons: this.setup(this.modelValue),
+        index: 0
       }
     },
     methods: {
@@ -63,12 +57,13 @@
       closeOptions (index) {
         this.$refs['popups-' + index][0].hide()
       },
-      submit () {
-        this.$emit('input', this.buttons.slice().filter(_ => _))
+      updateButton (index, control) {
+        this.buttons[index] = control
+        this.$emit('update:modelValue', this.buttons.slice().filter(_ => _))
       }
     },
     watch: {
-      value: function (value) {
+      modelValue: function (value) {
         this.buttons = this.setup(value)
       }
     }
